Revert "Migrate process_group to phi [fluid_ops]  (#67243)"
This reverts commit c2d2b669d816ba7e114ec48e3a2013aab98313c2.
