+++ paddle/fluid/distributed/collective/custom_ccl_tools.h
@@ -22,11 +22,13 @@
 
 #include <string>
 
+#include "paddle/fluid/framework/data_type.h"
+#include "paddle/fluid/framework/variable.h"
+#include "paddle/fluid/platform/collective_helper.h"
+#include "paddle/fluid/platform/enforce.h"
 #include "paddle/phi/backends/device_guard.h"
 #include "paddle/phi/backends/device_manager.h"
 #include "paddle/phi/core/distributed/types.h"
-#include "paddle/phi/core/enforce.h"
-#include "paddle/phi/core/platform/collective_helper.h"
 #include "paddle/phi/core/platform/device_context.h"
 
 namespace paddle {
