+++ paddle/fluid/distributed/collective/process_group_bkcl.cc
@@ -17,13 +17,13 @@
 #include "paddle/common/errors.h"
 #include "paddle/fluid/distributed/collective/bkcl_tools.h"
 #include "paddle/fluid/distributed/collective/common.h"
+#include "paddle/fluid/framework/convert_utils.h"
+#include "paddle/fluid/platform/device/xpu/bkcl_helper.h"
 #include "paddle/phi/api/lib/utils/allocator.h"
 #include "paddle/phi/core/device_context.h"
 #include "paddle/phi/core/distributed/check/static_check.h"
 #include "paddle/phi/core/distributed/comm_context_manager.h"
 #include "paddle/phi/core/enforce.h"
-#include "paddle/phi/core/memory/allocation/allocator_facade.h"
-#include "paddle/phi/core/platform/device/xpu/bkcl_helper.h"
 #include "paddle/phi/core/platform/device/xpu/xpu_info.h"
 
 namespace paddle {
