+++ paddle/fluid/distributed/collective/process_group_gloo.h
@@ -20,10 +20,13 @@
 
 #include "paddle/fluid/distributed/collective/process_group.h"
 #include "paddle/fluid/distributed/collective/process_group_without_stream.h"
-#include "paddle/phi/backends/context_pool.h"
 #include "paddle/phi/core/distributed/gloo_comm_context.h"
 #include "paddle/phi/core/distributed/store/store.h"
 #include "paddle/phi/core/distributed/store/tcp_store.h"
+
+#ifdef PADDLE_WITH_GLOO
+#include "paddle/fluid/framework/fleet/gloo_wrapper.h"
+#endif
 
 namespace paddle {
 namespace distributed {
@@ -55,13 +58,33 @@ class ProcessGroupGloo : public ProcessGroupWithoutStream {
 
     ~GlooStore() = default;
 
-    std::vector<char> get(const std::string& key) override;
-    void wait(const std::vector<std::string>& keys) override;
+    std::vector<char> get(const std::string& key) override {
+      VLOG(3) << "GlooStore::get";
+      auto value = _store->get(key);
+      return std::vector<char>(value.begin(), value.end());
+    }
 
-    void set(const std::string& key, const std::vector<char>& value) override;
+    void wait(const std::vector<std::string>& keys) override {
+      VLOG(3) << "GlooStore::wait";
+      for (auto& key : keys) {
+        _store->wait(key);
+      }
+    }
+
+    void set(const std::string& key, const std::vector<char>& value) override {
+      VLOG(3) << "GlooStore::set";
+      std::vector<uint8_t> tmp(value.begin(), value.end());
+      _store->set(key, tmp);
+    }
 
     void wait(const std::vector<std::string>& keys,
-              const std::chrono::milliseconds& timeout) override;
+              const std::chrono::milliseconds& timeout) override {
+      VLOG(3) << "GlooStore::wait";
+      for (auto& key : keys) {
+        _store->wait(key);
+      }
+      // wait(keys);
+    }
 
    protected:
     std::shared_ptr<phi::distributed::Store> _store;
