+++ paddle/fluid/distributed/collective/process_group_nccl.cc
@@ -15,6 +15,7 @@
 #include "paddle/fluid/distributed/collective/process_group_nccl.h"
 #include "paddle/common/flags.h"
 #include "paddle/fluid/distributed/collective/common.h"
+#include "paddle/fluid/platform/device/gpu/nccl_helper.h"
 #include "paddle/phi/api/lib/utils/allocator.h"
 #include "paddle/phi/backends/gpu/gpu_info.h"
 #include "paddle/phi/common/memory_utils.h"
@@ -27,7 +28,6 @@
 #include "paddle/phi/core/distributed/utils.h"
 #include "paddle/phi/core/enforce.h"
 #include "paddle/phi/core/platform/cuda_device_guard.h"
-#include "paddle/phi/core/platform/device/gpu/nccl_helper.h"
 #include "paddle/phi/core/utils/data_type.h"
 
 COMMON_DECLARE_bool(benchmark);
