+++ paddle/phi/core/distributed/collective/CMakeLists.txt
@@ -1,3 +1,3 @@
-set(COLLECTIVE_SRCS process_group.cc)
+set(COLLECTIVE_COMMON_SRCS process_group.cc)
 
-collect_srcs(core_srcs SRCS ${COLLECTIVE_SRCS})
+collect_srcs(core_srcs SRCS ${COLLECTIVE_COMMON_SRCS})
