--- paddle/phi/core/platform/device/gpu/nccl_helper.h
@@ -1,43 +1,0 @@
-// Copyright (c) 2024 PaddlePaddle Authors. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#pragma once
-
-#if defined(PADDLE_WITH_NCCL) || defined(PADDLE_WITH_RCCL)
-#include <stdio.h>
-
-#include <memory>
-#include <string>
-#include <thread>  // NOLINT
-#include <typeindex>
-#include <unordered_map>
-#include <vector>
-
-#include "paddle/phi/core/platform/collective_helper.h"
-#ifdef PADDLE_WITH_NCCL
-#include "paddle/phi/backends/dynload/nccl.h"
-#endif
-#ifdef PADDLE_WITH_RCCL
-#include "paddle/phi/backends/dynload/rccl.h"
-#endif
-#include "paddle/phi/common/bfloat16.h"
-#include "paddle/phi/common/float16.h"
-#include "paddle/phi/common/memory_utils.h"
-#include "paddle/phi/core/enforce.h"
-#include "paddle/phi/core/platform/device/gpu/gpu_dnn.h"
-
-namespace paddle {
-namespace platform {}  // namespace platform
-}  // namespace paddle
-#endif
